pyproject.toml
[project]
name = "fastmcp-plan-agent"
version = "0.1.0"
description = "FastMCP agentic MCP server for Oracle EPM Planning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Planning Team"}
]

dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "python-docx>=1.1.0",
    "anthropic>=0.34.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
planning-fastmcp-stdio = "cli.fastmcp_stdio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["planning_agent", "cli", "web"]

